[apps][ugdocu] Add C++ Generator + Cleanup
Now it is possible to generate commented (documented) C++ dummy code
from the registered classes and functions.
These C++ files can then be parsed by Doxygen to produce a nice docu for
the registered functionality of ug4.
This Doxygen-generated docu can then be integrated into the existing ug4
docu.

Also cleaned up the rest of the app and moved it into its own namespace.
Generation of the legacy HTML is still possible.